+++ cmd/run.go
@@ -4,6 +4,7 @@ import (
 	"github.com/robocorp/rcc/cloud"
 	"github.com/robocorp/rcc/common"
 	"github.com/robocorp/rcc/conda"
+	"github.com/robocorp/rcc/journal"
 	"github.com/robocorp/rcc/operations"
 	"github.com/robocorp/rcc/xviper"
 
@@ -24,6 +25,7 @@ var runCmd = &cobra.Command{
 in your own machine.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		defer conda.RemoveCurrentTemp()
+		defer journal.BuildEventStats("robot")
 		if common.DebugFlag {
 			defer common.Stopwatch("Task run lasted").Report()
 		}