+++ htfs/virtual.go
@@ -121,6 +121,7 @@ func (it *virtual) Restore(blueprint, client, tag []byte) (string, error) {
 	common.Timeline("holotree restore done (virtual)")
 	defer common.Timeline("- dirty %d/%d", score.dirty, score.total)
 	common.Debug("Holotree dirty workload: %d/%d\n", score.dirty, score.total)
+	journal.CurrentBuildEvent().Dirty(score.Dirtyness())
 	fs.Controller = string(client)
 	fs.Space = string(tag)
 	err = fs.SaveAs(metafile)